export is finish

+++ env.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   env.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abdelilah <abdelilah@student.42.fr>        +#+  +:+       +#+        */
+/*   By: amousaid <amousaid@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 12:55:25 by abdelilah         #+#    #+#             */
-/*   Updated: 2024/05/28 12:56:44 by abdelilah        ###   ########.fr       */
+/*   Updated: 2024/06/13 22:10:22 by amousaid         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "include/minishell.h"
 
-size_t find_env(char *str, t_minishell *mini)
+int find_env(char *str, t_minishell *mini)
 {
-    size_t i;
+    int i;
     size_t len;
     
     i = 0;
